# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "My Mechanic Shop API"
  description: "API for customers, mechanics, service tickets, and inventory."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login customer"
      description: "Validates customer credentials and returns an authentication token."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              status: "success"
              message: "Successfully logged in"
              auth_token: "jwt-token-here"
  /customers:
    get:
      tags:
        - Customers
      summary: "Get all customers"
      description: "Retrieves all customers."
      responses:
        200:
          description: "Retrieved customers successfully"
          schema:
            $ref: "#/definitions/AllCustomers"

    post:
      tags:
        - Customers
      summary: "Create customer"
      description: "Creates/registers a new customer."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/Customer"
  /customers/{customer_id}:
    parameters:
      - name: customer_id
        in: path
        required: true
        type: integer

    get:
      tags:
        - Customers
      summary: "Get customer by ID"
      description: "Retrieves one customer by ID."
      responses:
        200:
          description: "Customer found"
          schema:
            $ref: "#/definitions/Customer"

    put:
      tags:
        - Customers
      summary: "Update customer"
      description: "Updates customer information. Requires token authentication."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/Customer"

    delete:
      tags:
        - Customers
      summary: "Delete customer"
      description: "Deletes a customer by ID."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
...
definitions:
  LoginCredentials:
    type: object
    properties:
      email:
        type: string
        example: "test@email.com"
      password:
        type: string
        example: "test"
    required:
      - email
      - password
  LoginResponse:
    type: object
    properties:
      status:
        type: string
        example: "success"
      message:
        type: string
        example: "Successfully logged in"
      auth_token:
        type: string
        example: "jwt-token-here"
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: "Operation completed successfully"
  CreateCustomerPayload:
    type: object
    properties:
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "jd@email.com"
      phone_number:
        type: string
        example: "999-999-9999"
      password:
        type: string
        example: "123"
    required:
      - name
      - email
      - phone_number
      - password
  UpdateCustomerPayload:
    type: object
    properties:
      name:
        type: string
        example: "Peter"
      email:
        type: string
        example: "peter@email.com"
      phone_number:
        type: string
        example: "111-222-3333"
      password:
        type: string
        example: "newpassword"
  Customer:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "jd@email.com"
      phone_number:
        type: string
        example: "999-999-9999"
      password:
        type: string
        example: "123"
  AllCustomers:
    type: array
    items:
      $ref: "#/definitions/Customer"
...